[PW-2788] Add payment methods to the plugin and adjust card payment handler (#45)
* [PW-2788] Creating iDeal PM

* [PW-2788] Adjusting (de)activate PM functions

* [PW-2788] Adding extra payment methods

* [PW-2788] phpcs fixes

* [PW-2788] phpcs fixes

* [PW-2788] Removing unused method implementations
