+++ src/Handlers/CardsPaymentMethodHandler.php
@@ -1,6 +1,5 @@
-<?php
+<?php declare(strict_types=1);
 
-declare(strict_types=1);
 /**
  *                       ######
  *                       ######
@@ -26,510 +25,18 @@ declare(strict_types=1);
 
 namespace Adyen\Shopware\Handlers;
 
-use Adyen\AdyenException;
-use Adyen\Service\Builder\Address;
-use Adyen\Service\Builder\Browser;
-use Adyen\Service\Builder\Customer;
-use Adyen\Service\Builder\Payment;
-use Adyen\Service\Validator\CheckoutStateDataValidator;
-use Adyen\Shopware\Exception\PaymentException;
-use Adyen\Shopware\Service\PaymentStateDataService;
-use Adyen\Shopware\Storefront\Controller\RedirectResultController;
-use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
-use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\AsynchronousPaymentHandlerInterface;
 use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
-use Shopware\Core\Checkout\Payment\Exception\AsyncPaymentFinalizeException;
-use Shopware\Core\Checkout\Payment\Exception\AsyncPaymentProcessException;
+use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\AsynchronousPaymentHandlerInterface;
 use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Adyen\Shopware\Service\CheckoutService;
-use Adyen\Util\Currency;
-use Adyen\Shopware\Service\ConfigurationService;
-use Psr\Log\LoggerInterface;
-use Adyen\Shopware\Service\Repository\SalesChannelRepository;
-use \Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
-class CardsPaymentMethodHandler implements AsynchronousPaymentHandlerInterface
+class CardsPaymentMethodHandler extends AbstractPaymentMethodHandler implements AsynchronousPaymentHandlerInterface
 {
-    /**
-     * @var CheckoutService
-     */
-    protected $checkoutService;
 
-    /**
-     * @var Browser
-     */
-    protected $browserBuilder;
-
-    /**
-     * @var Address
-     */
-    protected $addressBuilder;
-
-    /**
-     * @var Payment
-     */
-    protected $paymentBuilder;
-
-    /**
-     * @var Currency
-     */
-    protected $currency;
-
-    /**
-     * @var ConfigurationService
-     */
-    protected $configurationService;
-
-    /**
-     * @var Customer
-     */
-    protected $customerBuilder;
-
-    /**
-     * @var CheckoutStateDataValidator
-     */
-    protected $checkoutStateDataValidator;
-
-    /**
-     * @var PaymentStateDataService
-     */
-    protected $paymentStateDataService;
-
-    /**
-     * @var LoggerInterface
-     */
-    protected $logger;
-
-    /**
-     * @var SalesChannelRepository
-     */
-    protected $salesChannelRepository;
-
-    /**
-     * @var PaymentResponseHandler
-     */
-    protected $paymentResponseHandler;
-
-    /**
-     * @var ResultHandler
-     */
-    protected $resultHandler;
-
-    /**
-     * @var OrderTransactionStateHandler
-     */
-    protected $orderTransactionStateHandler;
-
-    /**
-     * @var RouterInterface
-     */
-    protected $router;
-
-    /**
-     * @var CsrfTokenManagerInterface
-     */
-    protected $csrfTokenManager;
-
-    /**
-     * CardsPaymentMethodHandler constructor.
-     *
-     * @param ConfigurationService $configurationService
-     * @param CheckoutService $checkoutService
-     * @param Browser $browserBuilder
-     * @param Address $addressBuilder
-     * @param Payment $paymentBuilder
-     * @param Currency $currency
-     * @param Customer $customerBuilder
-     * @param CheckoutStateDataValidator $checkoutStateDataValidator
-     * @param PaymentStateDataService $paymentStateDataService
-     * @param SalesChannelRepository $salesChannelRepository
-     * @param PaymentResponseHandler $paymentResponseHandler
-     * @param ResultHandler $resultHandler
-     * @param OrderTransactionStateHandler $orderTransactionStateHandler
-     * @param RouterInterface $router
-     * @param LoggerInterface $logger
-     */
-    public function __construct(
-        ConfigurationService $configurationService,
-        CheckoutService $checkoutService,
-        Browser $browserBuilder,
-        Address $addressBuilder,
-        Payment $paymentBuilder,
-        Currency $currency,
-        Customer $customerBuilder,
-        CheckoutStateDataValidator $checkoutStateDataValidator,
-        PaymentStateDataService $paymentStateDataService,
-        SalesChannelRepository $salesChannelRepository,
-        PaymentResponseHandler $paymentResponseHandler,
-        ResultHandler $resultHandler,
-        OrderTransactionStateHandler $orderTransactionStateHandler,
-        RouterInterface $router,
-        CsrfTokenManagerInterface $csrfTokenManager,
-        LoggerInterface $logger
-    ) {
-        $this->checkoutService = $checkoutService;
-        $this->browserBuilder = $browserBuilder;
-        $this->addressBuilder = $addressBuilder;
-        $this->currency = $currency;
-        $this->configurationService = $configurationService;
-        $this->customerBuilder = $customerBuilder;
-        $this->paymentBuilder = $paymentBuilder;
-        $this->checkoutStateDataValidator = $checkoutStateDataValidator;
-        $this->paymentStateDataService = $paymentStateDataService;
-        $this->salesChannelRepository = $salesChannelRepository;
-        $this->paymentResponseHandler = $paymentResponseHandler;
-        $this->resultHandler = $resultHandler;
-        $this->logger = $logger;
-        $this->orderTransactionStateHandler = $orderTransactionStateHandler;
-        $this->router = $router;
-        $this->csrfTokenManager = $csrfTokenManager;
-    }
-
-    /**
-     * @return string
-     */
-    public function getAdyenPaymentMethodId(): string
+    public static function getPaymentMethodCode()
     {
         return 'scheme';
-    }
-
-    /**
-     * @param AsyncPaymentTransactionStruct $transaction
-     * @param RequestDataBag $dataBag
-     * @param SalesChannelContext $salesChannelContext
-     * @return RedirectResponse
-     */
-    public function pay(
-        AsyncPaymentTransactionStruct $transaction,
-        RequestDataBag $dataBag,
-        SalesChannelContext $salesChannelContext
-    ): RedirectResponse {
-        try {
-            $request = $this->preparePaymentsRequest($salesChannelContext, $transaction);
-        } catch (Exception $exception) {
-            $message = sprintf(
-                "There was an error with the payment method. Order number: %s Missing data: %s",
-                $transaction->getOrder()->getOrderNumber(),
-                $exception->getMessage()
-            );
-            $this->logger->error($message);
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                $message
-            );
-        }
-
-        try {
-            $response = $this->checkoutService->payments($request);
-        } catch (AdyenException $exception) {
-            $message = sprintf(
-                "There was an error with the /payments request. Order number %s: %s",
-                $transaction->getOrder()->getOrderNumber(),
-                $exception->getMessage()
-            );
-
-            $this->logger->error($message);
-
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                $message
-            );
-        }
-
-        $orderNumber = $transaction->getOrder()->getOrderNumber();
-
-        if (empty($orderNumber)) {
-            $message = 'Order number is missing';
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                $message
-            );
-        }
-
-        $result = $this->paymentResponseHandler->handlePaymentResponse($response, $orderNumber, $salesChannelContext);
-
-        try {
-            $this->paymentResponseHandler->handleShopwareApis($transaction, $salesChannelContext, $result);
-        } catch (PaymentException $exception) {
-            // Cancel payment in shopware
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                $exception->getMessage()
-            );
-        }
-
-        // Payment had no error, continue the process
-        return new RedirectResponse($this->getAdyenReturnUrl($transaction->getReturnUrl()));
-    }
-
-    /**
-     * @param AsyncPaymentTransactionStruct $transaction
-     * @param Request $request
-     * @param SalesChannelContext $salesChannelContext
-     */
-    public function finalize(
-        AsyncPaymentTransactionStruct $transaction,
-        Request $request,
-        SalesChannelContext $salesChannelContext
-    ): void {
-        try {
-            $this->resultHandler->processResult($transaction, $request, $salesChannelContext);
-        } catch (PaymentException $exception) {
-            throw new AsyncPaymentFinalizeException(
-                $transaction->getOrderTransaction()->getId(),
-                $exception->getMessage()
-            );
-        }
-    }
-
-    //TODO move to external object or outsource to lib
-
-    /**
-     * @param string $address
-     * @return array
-     */
-    public function splitStreetAddressHouseNumber(string $address): array
-    {
-        return [
-            'street' => $address,
-            'houseNumber' => 'N/A'
-        ];
-    }
-
-    /**
-     * @param SalesChannelContext $salesChannelContext
-     * @param AsyncPaymentTransactionStruct $transaction
-     * @return array
-     */
-    public function preparePaymentsRequest(
-        SalesChannelContext $salesChannelContext,
-        AsyncPaymentTransactionStruct $transaction
-    ) {
-        //Get state.data using the context token
-        $stateData = $this->paymentStateDataService->getPaymentStateDataFromContextToken(
-            $salesChannelContext->getToken()
-        );
-        $request = json_decode($stateData->getStateData(), true);
-
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                'Invalid payment state data.'
-            );
-        }
-
-        if (!empty($request['additionalData'])) {
-            $stateDataAdditionalData = $request['additionalData'];
-        }
-
-        //Validate state.data for payment and build request object
-        $request = $this->checkoutStateDataValidator->getValidatedAdditionalData($request);
-
-        //Setting browser info if not present in statedata
-        if (empty($request['browserInfo']['acceptHeader'])) {
-            $acceptHeader = $_SERVER['HTTP_ACCEPT'];
-        } else {
-            $acceptHeader = $request['browserInfo']['acceptHeader'];
-        }
-        if (empty($request['browserInfo']['userAgent'])) {
-            $userAgent = $_SERVER['HTTP_USER_AGENT'];
-        } else {
-            $userAgent = $request['browserInfo']['userAgent'];
-        }
-
-        //Setting delivery address info if not present in statedata
-        if (empty($request['deliveryAddress'])) {
-            if ($salesChannelContext->getShippingLocation()->getAddress()->getCountryState()) {
-                $shippingState = $salesChannelContext->getShippingLocation()
-                    ->getAddress()->getCountryState()->getShortCode();
-            } else {
-                $shippingState = '';
-            }
-
-            $shippingStreetAddress = $this->splitStreetAddressHouseNumber(
-                $salesChannelContext->getShippingLocation()->getAddress()->getStreet()
-            );
-            $request = $this->addressBuilder->buildDeliveryAddress(
-                $shippingStreetAddress['street'],
-                $shippingStreetAddress['houseNumber'],
-                $salesChannelContext->getShippingLocation()->getAddress()->getZipcode(),
-                $salesChannelContext->getShippingLocation()->getAddress()->getCity(),
-                $shippingState,
-                $salesChannelContext->getShippingLocation()->getAddress()->getCountry()->getIso(),
-                $request
-            );
-        }
-
-        //Setting billing address info if not present in statedata
-        if (empty($request['billingAddress'])) {
-            if ($salesChannelContext->getCustomer()->getActiveBillingAddress()->getCountryState()) {
-                $billingState = $salesChannelContext->getCustomer()
-                    ->getActiveBillingAddress()->getCountryState()->getShortCode();
-            } else {
-                $billingState = '';
-            }
-
-            $billingStreetAddress = $this->splitStreetAddressHouseNumber(
-                $salesChannelContext->getCustomer()->getActiveBillingAddress()->getStreet()
-            );
-            $request = $this->addressBuilder->buildBillingAddress(
-                $billingStreetAddress['street'],
-                $billingStreetAddress['houseNumber'],
-                $salesChannelContext->getCustomer()->getActiveBillingAddress()->getZipcode(),
-                $salesChannelContext->getCustomer()->getActiveBillingAddress()->getCity(),
-                $billingState,
-                $salesChannelContext->getCustomer()->getActiveBillingAddress()->getCountry()->getIso(),
-                $request
-            );
-        }
-
-        //Setting customer data if not present in statedata
-        if (empty($request['shopperName'])) {
-            $shopperFirstName = $salesChannelContext->getCustomer()->getFirstName();
-            $shopperLastName = $salesChannelContext->getCustomer()->getFirstName();
-        } else {
-            $shopperFirstName = $request['shopperName']['firstName'];
-            $shopperLastName = $request['shopperName']['lastName'];
-        }
-
-        if (empty($request['shopperEmail'])) {
-            $shopperEmail = $salesChannelContext->getCustomer()->getEmail();
-        } else {
-            $shopperEmail = $request['shopperEmail'];
-        }
-
-        if (empty($request['paymentMethod']['personalDetails']['telephoneNumber'])) {
-            $shopperPhone = $salesChannelContext->getShippingLocation()->getAddress()->getPhoneNumber();
-        } else {
-            $shopperPhone = $request['paymentMethod']['personalDetails']['telephoneNumber'];
-        }
-
-        if (empty($request['paymentMethod']['personalDetails']['dateOfBirth'])) {
-            if ($salesChannelContext->getCustomer()->getBirthday()) {
-                $shopperDob = $salesChannelContext->getCustomer()->getBirthday()->format('dd-mm-yyyy');
-            } else {
-                $shopperDob = '';
-            }
-        } else {
-            $shopperDob = $request['paymentMethod']['personalDetails']['dateOfBirth'];
-        }
-
-        if (empty($request['shopperLocale'])) {
-            $shopperLocale = $this->salesChannelRepository->getSalesChannelAssocLocale($salesChannelContext)
-                ->getLanguage()->getLocale()->getCode();
-        } else {
-            $shopperLocale = $request['shopperLocale'];
-        }
-
-        if (empty($request['shopperIP'])) {
-            $shopperIp = $salesChannelContext->getCustomer()->getRemoteAddress();
-        } else {
-            $shopperIp = $request['shopperIP'];
-        }
-
-        if (empty($request['shopperReference'])) {
-            $shopperReference = $salesChannelContext->getCustomer()->getId();
-        } else {
-            $shopperReference = $request['shopperReference'];
-        }
-
-        if (empty($request['countryCode'])) {
-            $countryCode = $salesChannelContext->getCustomer()->getActiveBillingAddress()->getCountry()->getIso();
-        } else {
-            $countryCode = $request['countryCode'];
-        }
-
-        $request = $this->browserBuilder->buildBrowserData(
-            $userAgent,
-            $acceptHeader,
-            $request['browserInfo']['screenWidth'],
-            $request['browserInfo']['screenHeight'],
-            $request['browserInfo']['colorDepth'],
-            $request['browserInfo']['timeZoneOffset'],
-            $request['browserInfo']['language'],
-            $request['browserInfo']['javaEnabled'],
-            $request
-        );
-
-        $request = $this->customerBuilder->buildCustomerData(
-            false,
-            $shopperEmail,
-            $shopperPhone,
-            '',
-            $shopperDob,
-            $shopperFirstName,
-            $shopperLastName,
-            $countryCode,
-            $shopperLocale,
-            $shopperIp,
-            $shopperReference,
-            $request
-        );
-
-        //Building payment data
-        $request = $this->paymentBuilder->buildPaymentData(
-            $salesChannelContext->getCurrency()->getIsoCode(),
-            $this->currency->sanitize(
-                $transaction->getOrder()->getPrice()->getTotalPrice(),
-                $salesChannelContext->getCurrency()->getIsoCode()
-            ),
-            $transaction->getOrder()->getOrderNumber(),
-            $this->configurationService->getMerchantAccount(),
-            $this->getAdyenReturnUrl($transaction->getReturnUrl()),
-            $request
-        );
-
-        //Setting info from statedata additionalData if present
-        if (!empty($stateDataAdditionalData['origin'])) {
-            $request['origin'] = $stateDataAdditionalData['origin'];
-        } else {
-            $request['origin'] = $this->salesChannelRepository->getSalesChannelUrl($salesChannelContext);
-        }
-
-        $request['additionalData']['allow3DS2'] = true;
-
-        $request['channel'] = 'web';
-
-        //Remove the used state.data
-        $this->paymentStateDataService->deletePaymentStateData($stateData);
-
-        return $request;
-    }
-
-    /**
-     * Creates the Adyen Redirect Result URL with the same query as the original return URL
-     * Fixes the CSRF validation bug: https://issues.shopware.com/issues/NEXT-6356
-     *
-     * @param $returnUrl
-     * @return string
-     * @throws PaymentException
-     */
-    private function getAdyenReturnUrl($returnUrl)
-    {
-        // Parse the original return URL to retrieve the query parameters
-        $returnUrlQuery = parse_url($returnUrl, PHP_URL_QUERY);
-
-        // In case the URL is malformed it cannot be parsed
-        if (false === $returnUrlQuery) {
-            throw new PaymentException('Return URL is malformed');
-        }
-
-        // Generate the custom Adyen endpoint to receive the redirect from the issuer page
-        $adyenReturnUrl = $this->router->generate(
-            'adyen_redirect_result',
-            [
-                RedirectResultController::CSRF_TOKEN => $this->csrfTokenManager->getToken(
-                    'payment.finalize.transaction'
-                )->getValue()
-            ],
-            RouterInterface::ABSOLUTE_URL
-        );
-
-        // Create the adyen redirect result URL with the same query as the original return URL
-        return $adyenReturnUrl . '&' . $returnUrlQuery;
     }
 }
